# Faster mutex implementation (optionally reentrant)