--- a/src/mutex.cr
+++ b/src/mutex.cr
@@ -2,35 +2,71 @@ require "crystal/spin_lock"
 
 # A fiber-safe mutex.
 class Mutex
+  @state = Atomic(Int32).new(0)
   @mutex_fiber : Fiber?
   @lock_count = 0
   @queue = Deque(Fiber).new
+  @queue_count = Atomic(Int32).new(0)
   @lock = Crystal::SpinLock.new
 
-  def lock
-    @lock.lock
-    mutex_fiber = @mutex_fiber
-    current_fiber = Fiber.current
+  def initialize(*, @reentrant = false)
+  end
 
-    if !mutex_fiber
-      @mutex_fiber = current_fiber
-      @lock.unlock
-    elsif mutex_fiber == current_fiber
-      @lock_count += 1 # recursive lock
-      @lock.unlock
-    else
-      @queue << current_fiber
-      @lock.unlock
-      Crystal::Scheduler.reschedule
+  @[AlwaysInline]
+  def lock
+    if @state.swap(1) == 0
+      @mutex_fiber = Fiber.current if @reentrant
+      return
     end
 
+    if @reentrant && @mutex_fiber == Fiber.current
+      @lock_count += 1
+      return
+    end
+
+    lock_slow
     nil
   end
 
-  def unlock
-    @lock.lock
+  @[NoInline]
+  private def lock_slow
+    loop do
+      break if try_lock
 
-    begin
+      @lock.sync do
+        @queue_count.add(1)
+        if @state.get == 0
+          if @state.swap(1) == 0
+            @queue_count.add(-1)
+            @mutex_fiber = Fiber.current
+            return
+          end
+        end
+
+        @queue.push Fiber.current
+      end
+      Crystal::Scheduler.reschedule
+    end
+
+    @mutex_fiber = Fiber.current
+    nil
+  end
+
+  private def try_lock
+    i = 1000
+    while @state.swap(1) != 0
+      while @state.get != 0
+        Intrinsics.pause
+        i &-= 1
+        return false if i == 0
+      end
+    end
+
+    true
+  end
+
+  def unlock
+    if @reentrant
       unless @mutex_fiber == Fiber.current
         raise "Attempt to unlock a mutex which is not locked"
       end
@@ -40,15 +76,26 @@ class Mutex
         return
       end
 
-      if fiber = @queue.try &.shift?
-        @mutex_fiber = fiber
-        fiber.enqueue
-      else
-        @mutex_fiber = nil
-      end
-    ensure
-      @lock.unlock
+      @mutex_fiber = nil
     end
+
+    @state.lazy_set(0)
+
+    if @queue_count.get == 0
+      return
+    end
+
+    fiber = nil
+    @lock.sync do
+      if @queue_count.get == 0
+        return
+      end
+
+      if fiber = @queue.shift?
+        @queue_count.add(-1)
+      end
+    end
+    fiber.enqueue if fiber
 
     nil
   end