@ysbaddaden 2019-10-08
Ah, thanks for the clarification. It makes sense now. Yes, spinning is always a good idea :)