@RX14 2019-10-21
I really don't think having non-reentrant (recursive) be the default is a good idea. What happens if you accidentally recurse? A deadlock? Corruption? The only acceptable state is an exception, or having a recursive lock. There may be an unsafe option to make the lock much faster for the cost of having to be extra careful not to get it wrong, but safety *must* be the default.