@asterite 2019-10-21
I agree. Crystal should always put safety before performance.