@asterite 2019-10-29
I think this will mostly hit users in a silent way when doing stuff like:

```crystal
# called from multiple functions
def helper_function
  @mutex.synchronize do
    # some data
  end
end

def main_function
  @mutex.synchronize do
    do_something
    helper_function # Oops, deadlock, not easy to debug
  end
end
```