@jhass 2019-10-29
Would it cost too much performance to require an explicit lock owner akin to Java's `synchronized`? It shouldn't be more than a pointer comparison and an , right? Something like

```cr
def lock(owner)
  raise "Already locked" if @current_owner = owner.object_id
  real_lock
  @current_owner = owner.object_id
end
```

Or actually a CAS operation for that.

Then Ary's example would raise.