@bcardiff 2019-10-29
But if you are using a lock that is hold you want to wait for it to be free, not fail immediately. Maybe a reasonable timeout on debug mode can help, but I am not sure I even like that alternative.

Or you mean that only for non-reentrant mutex to signal a hey: you might need a reentrant mutex?