@RX14 2019-11-22
Still:

> Is a benchmark of the non-reentrant vs reentrant available?