@lribeiro 2019-11-23
That would be my preference as well, safe by default, fast by need.
`.new(reentrant: true)` seems clear