@ysbaddaden 2019-11-25
Yes, and that's why I insisted above on having all 3 states:
1. check by default: raise on deadlock + wrong fiber unlocking (safe);
2. reentrant option (may still deadlock when skipping an unlock, but otherwise safe);
3. disable checks option: full speed (unsafe: can deadlock + any fiber can unlock).

I.e. similar to pthreads options but checked by default. It prevents deadlocks by default and reports where a deadlock/invalid unlock occurred —without having to reproduce in GDB. It keeps some flexibility with reentrant, and it won't impact performance when checks are disabled (we already check for reentrant).